Generalize TCP over IP port number implementation
Resolves #671 (Generalize TCP over IP port number implementation).
